setting up pytest and cleaning up deprecated datetime.utcnow

+++ src/purchase/functions.py
@@ -1,4 +1,5 @@
 from src.extensions import db
+from src.models.history import History
 from src.models.inventory import Inventory
 from src.models.purchase_order import PurchaseOrder
 from src.purchase.schemas import PurchaseOrderStates
@@ -24,30 +25,30 @@ def create_inventory(quantity, **data):
     return new_inventory
 
 
-def create_purchase(author_id,
-                    item_id,
-                    price,
-                    currency_id,
-                    to_rial_rate,
-                    make_unique,
-                    provider_id,
-                    quantity=1,
-                    warehouse_id=None,
-                    state='Ordered',
-                    creation_date=None,
-                    update_date=None
-                    ):
+def create_purchase(
+    user_id,
+    item_id,
+    price,
+    currency_id,
+    base_currency_rate,
+    make_unique,
+    provider_id,
+    quantity=1,
+    warehouse_id=None,
+    state='Ordered',
+    creation_date=None,
+    update_date=None
+):
     try:
         new_purchase = PurchaseOrder(
             quantity=quantity,
             provider_id=provider_id,
-            author_id=author_id,
             creation_date=creation_date,
             state=state,
             update_date=update_date,
             make_unique=make_unique,
             price=price,
-            to_rial_rate=to_rial_rate,
+            base_currency_rate=base_currency_rate,
             item_id=item_id,
             warehouse_id=warehouse_id,
             currency_id=currency_id
@@ -64,27 +65,27 @@ def create_purchase(author_id,
         print('[create_purchase]', err)
         return {"err": "Somthing went wrong! please try again later"}, 500
 
-    try:
-        purchase_to_invetories(
-            item_id=item_id,
-            author_id=author_id,
-            price=price,
-            currency_id=currency_id,
-            to_rial_rate=to_rial_rate,
-            purchase_id=new_purchase.id,
-            quantity=quantity,
-            make_unique=make_unique,
-            provider_id=provider_id,
-            warehouse_id=warehouse_id,
-            creation_date=creation_date,
-            update_date=update_date
-        )
-    except Exception as err:
-        db.session.rollback()
-        db.session.delete(new_purchase)
-        db.session.commit()
-        print('[create_purchase]', err)
-        return {"err": "Something went wrong!, please try again later"}, 500
+    if state == PurchaseOrderStates.Received.value:
+        try:
+            purchase_to_invetories(
+                item_id=item_id,
+                price=price,
+                currency_id=currency_id,
+                base_currency_rate=base_currency_rate,
+                purchase_id=new_purchase.id,
+                quantity=quantity,
+                make_unique=make_unique,
+                provider_id=provider_id,
+                warehouse_id=warehouse_id,
+                creation_date=creation_date,
+                update_date=update_date
+            )
+        except Exception as err:
+            db.session.rollback()
+            db.session.delete(new_purchase)
+            db.session.commit()
+            print('[create_purchase]', err)
+            return {"err": "Something went wrong!, please try again later"}, 500
 
     return {"Purchase Order": new_purchase.json()}, 201
 
@@ -102,44 +103,49 @@ def update_purchase(id, **data):
     inventories = Inventory.query.where(Inventory.purchase_id == id).all()
     if len(inventories) != 0:
         print(inventories)
-        return {"err": "This action could not be completed! please contact admin"}, 500
+        return {
+            "err": "This action could not be completed! please contact admin"
+        }, 500
 
+    state = data.get('state', purchase.state)
+    warehouse_id = data.get('warehouse_id', purchase.warehouse_id)
+    if state == PurchaseOrderStates.Received.value and warehouse_id is None:
+        return {"err": "You cannot convert PO to Received without Warehouse!"}
     item_id = data.get('item_id', purchase.item_id)
-    author_id = data.get('author_id', purchase.author_id)
     price = data.get('price', purchase.price)
     currency_id = data.get('currency_id', purchase.currency_id)
-    to_rial_rate = data.get('to_rial_rate', purchase.to_rial_rate)
+    base_currency_rate = data.get(
+        'base_currency_rate', purchase.base_currency_rate)
     creation_date = data.get('creation_date', purchase.creation_date)
     update_date = data.get('update_date', None)
     quantity = data.get('quantity', purchase.quantity)
     provider_id = data.get('provider_id', purchase.provider_id)
     make_unique = data.get('make_unique', purchase.make_unique)
-    warehouse_id = data.get('warehouse_id', purchase.warehouse_id)
 
-    try:
-        purchase_to_invetories(
-            item_id=item_id,
-            author_id=author_id,
-            price=price,
-            currency_id=currency_id,
-            to_rial_rate=to_rial_rate,
-            purchase_id=purchase.id,
-            quantity=quantity,
-            make_unique=make_unique,
-            provider_id=provider_id,
-            warehouse_id=warehouse_id,
-            creation_date=creation_date,
-            update_date=update_date
-        )
-    except Exception as err:
-        print(['update_purchase'], err)
-        return {"err": "Somthing went wrong! please try again later"}
+    if state == PurchaseOrderStates.Received.value:
+        try:
+            purchase_to_invetories(
+                item_id=item_id,
+                price=price,
+                currency_id=currency_id,
+                base_currency_rate=base_currency_rate,
+                purchase_id=purchase.id,
+                quantity=quantity,
+                make_unique=make_unique,
+                provider_id=provider_id,
+                warehouse_id=warehouse_id,
+                creation_date=creation_date,
+                update_date=update_date
+            )
+        except Exception as err:
+            print(['update_purchase'], err)
+            return {"err": "Somthing went wrong! please try again later"}
 
-    purchase.state = 'Received'
+    purchase.state = state
     purchase.item_id = item_id
     purchase.price = price
     purchase.currency_id = currency_id
-    purchase.to_rial_rate = to_rial_rate
+    purchase.base_currency_rate = base_currency_rate
     purchase.creation_date = creation_date
     purchase.update_date = update_date
     purchase.quantity = quantity
